store in one slot

+++ contracts/oracles/libraries/IporRiskManagementOracleStorageTypes.sol
@@ -20,30 +20,30 @@ library IporRiskManagementOracleStorageTypes {
 
     struct BaseSpreadsAndFixedRateCapsStorage {
         /// @notice Timestamp of most recent indicators update
-        uint32 lastUpdateTimestamp;
+        uint256 lastUpdateTimestamp;
         /// @notice spread for 28 days pay fixed swap
-        int24 spread28dPayFixed;
+        int256 spread28dPayFixed;
         /// @notice spread for 28 days receive fixed swap
-        int24 spread28dReceiveFixed;
+        int256 spread28dReceiveFixed;
         /// @notice spread for 60 days pay fixed swap
-        int24 spread60dPayFixed;
+        int256 spread60dPayFixed;
         /// @notice spread for 60 days receive fixed swap
-        int24 spread60dReceiveFixed;
+        int256 spread60dReceiveFixed;
         /// @notice spread for 90 days pay fixed swap
-        int24 spread90dPayFixed;
+        int256 spread90dPayFixed;
         /// @notice spread for 90 days receive fixed swap
-        int24 spread90dReceiveFixed;
+        int256 spread90dReceiveFixed;
         /// @notice fixed rate cap for 28 days pay fixed swap
-        uint16 fixedRateCap28dPayFixed;
+        uint256 fixedRateCap28dPayFixed;
         /// @notice fixed rate cap for 28 days receive fixed swap
-        uint16 fixedRateCap28dReceiveFixed;
+        uint256 fixedRateCap28dReceiveFixed;
         /// @notice fixed rate cap for 60 days pay fixed swap
-        uint16 fixedRateCap60dPayFixed;
+        uint256 fixedRateCap60dPayFixed;
         /// @notice fixed rate cap for 60 days receive fixed swap
-        uint16 fixedRateCap60dReceiveFixed;
+        uint256 fixedRateCap60dReceiveFixed;
         /// @notice fixed rate cap for 90 days pay fixed swap
-        uint16 fixedRateCap90dPayFixed;
+        uint256 fixedRateCap90dPayFixed;
         /// @notice fixed rate cap for 90 days receive fixed swap
-        uint16 fixedRateCap90dReceiveFixed;
+        uint256 fixedRateCap90dReceiveFixed;
     }
 }
